# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * LexAI - AI Legal Assistant Platform API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type RegisterBodyRole =
  (typeof RegisterBodyRole)[keyof typeof RegisterBodyRole];

export const RegisterBodyRole = {
  NORMAL: "NORMAL",
  LAWYER: "LAWYER",
} as const;

export interface RegisterBody {
  email: string;
  firstName: string;
  lastName: string;
  password: string;
  role: RegisterBodyRole;
  language?: string;
}

export interface LoginBody {
  email: string;
  password: string;
}

export type UserRole = (typeof UserRole)[keyof typeof UserRole];

export const UserRole = {
  NORMAL: "NORMAL",
  LAWYER: "LAWYER",
} as const;

export type UserMembershipPlan =
  (typeof UserMembershipPlan)[keyof typeof UserMembershipPlan];

export const UserMembershipPlan = {
  FREE: "FREE",
  PREMIUM: "PREMIUM",
} as const;

export interface User {
  id: number;
  email: string;
  firstName: string;
  lastName: string;
  role: UserRole;
  avatarUrl?: string;
  language?: string;
  membershipPlan: UserMembershipPlan;
  createdAt?: string;
}

export interface AuthResponse {
  user: User;
  token?: string;
}

export interface UserProfile {
  id: number;
  email: string;
  firstName: string;
  lastName: string;
  role: string;
  avatarUrl?: string;
  language?: string;
  membershipPlan: string;
  bio?: string;
  phone?: string;
  createdAt?: string;
}

export interface UpdateProfileBody {
  firstName?: string;
  lastName?: string;
  language?: string;
  bio?: string;
  phone?: string;
  avatarUrl?: string;
}

export interface LawyerOnboardingBody {
  barNumber: string;
  yearsOfExperience: number;
  languages: string[];
  specialization: string;
  age: number;
  bio?: string;
  hourlyRate?: number;
}

export type LawyerProfileVerificationStatus =
  (typeof LawyerProfileVerificationStatus)[keyof typeof LawyerProfileVerificationStatus];

export const LawyerProfileVerificationStatus = {
  PENDING: "PENDING",
  APPROVED: "APPROVED",
  REJECTED: "REJECTED",
} as const;

export interface LawyerProfile {
  id: number;
  userId: number;
  barNumber: string;
  yearsOfExperience: number;
  languages: string[];
  specialization: string;
  age: number;
  verificationStatus: LawyerProfileVerificationStatus;
  bio?: string;
  hourlyRate?: number;
  rating?: number;
}

export interface UsageStats {
  documentsUploadedToday: number;
  aiQueriesUsedToday: number;
  documentsLimit: number;
  aiQueriesLimit: number;
  membershipPlan: string;
}

export interface ChatSession {
  id: number;
  title: string;
  documentId?: number;
  lastMessage?: string;
  messageCount: number;
  createdAt: string;
  updatedAt: string;
}

export type ChatMessageRole =
  (typeof ChatMessageRole)[keyof typeof ChatMessageRole];

export const ChatMessageRole = {
  user: "user",
  assistant: "assistant",
} as const;

export interface ChatMessage {
  id: number;
  sessionId: number;
  role: ChatMessageRole;
  content: string;
  createdAt: string;
}

export interface ChatSessionDetail {
  id: number;
  title: string;
  documentId?: number;
  messages: ChatMessage[];
  createdAt: string;
  updatedAt: string;
}

export interface CreateChatSessionBody {
  title: string;
  documentId?: number;
}

export interface SendMessageBody {
  content: string;
  documentId?: number;
}

export interface Document {
  id: number;
  userId: number;
  name: string;
  fileType: string;
  fileSize?: number;
  extractedText?: string;
  summary?: string;
  uploadedAt: string;
}

export interface UploadDocumentBody {
  name: string;
  fileType: string;
  fileSize?: number;
  content?: string;
}

export type AnalyzeDocumentBodyAction =
  (typeof AnalyzeDocumentBodyAction)[keyof typeof AnalyzeDocumentBodyAction];

export const AnalyzeDocumentBodyAction = {
  summarize: "summarize",
  simplify: "simplify",
  detect_risks: "detect_risks",
  translate: "translate",
  ask_question: "ask_question",
} as const;

export interface AnalyzeDocumentBody {
  action: AnalyzeDocumentBodyAction;
  question?: string;
  targetLanguage?: string;
}

export interface DocumentAnalysis {
  documentId: number;
  action: string;
  result: string;
  disclaimer?: string;
}

export interface LawyerCard {
  id: number;
  userId: number;
  firstName: string;
  lastName: string;
  avatarUrl?: string;
  specialization: string;
  languages: string[];
  yearsOfExperience: number;
  rating: number;
  reviewCount: number;
  hourlyRate?: number;
  verificationStatus?: string;
}

export interface LawyerDetail {
  id: number;
  userId: number;
  firstName: string;
  lastName: string;
  avatarUrl?: string;
  specialization: string;
  languages: string[];
  yearsOfExperience: number;
  rating: number;
  reviewCount: number;
  hourlyRate?: number;
  bio?: string;
  barNumber?: string;
  verificationStatus?: string;
  availability?: string[];
}

export interface Review {
  id: number;
  lawyerId: number;
  userId: number;
  userName?: string;
  rating: number;
  comment?: string;
  createdAt: string;
}

export interface AddReviewBody {
  rating: number;
  comment?: string;
}

export type LawyerStatsBySpecializationItem = {
  specialization: string;
  count: number;
};

export interface LawyerStats {
  totalLawyers: number;
  bySpecialization: LawyerStatsBySpecializationItem[];
  topRatedLawyers: LawyerCard[];
  averageRating: number;
}

export interface Conversation {
  id: number;
  lawyerId: number;
  lawyerName: string;
  lawyerAvatarUrl?: string;
  lastMessage?: string;
  lastMessageAt?: string;
  unreadCount: number;
}

export interface DirectMessage {
  id: number;
  conversationId: number;
  senderId: number;
  content: string;
  createdAt: string;
}

export interface ConversationDetail {
  id: number;
  lawyerId: number;
  lawyerName: string;
  lawyerAvatarUrl?: string;
  messages: DirectMessage[];
}

export interface SendDirectMessageBody {
  content: string;
}

export interface StartConversationBody {
  lawyerId: number;
  initialMessage?: string;
}

export type BookingStatus = (typeof BookingStatus)[keyof typeof BookingStatus];

export const BookingStatus = {
  PENDING: "PENDING",
  CONFIRMED: "CONFIRMED",
  CANCELLED: "CANCELLED",
  COMPLETED: "COMPLETED",
} as const;

export interface Booking {
  id: number;
  userId: number;
  lawyerId: number;
  lawyerName: string;
  lawyerAvatarUrl?: string;
  scheduledAt: string;
  duration: number;
  meetLink?: string;
  status: BookingStatus;
  notes?: string;
}

export interface CreateBookingBody {
  lawyerId: number;
  scheduledAt: string;
  duration: number;
  notes?: string;
}

export interface TimeSlot {
  startTime: string;
  endTime: string;
  available: boolean;
}

export type MembershipPlan =
  (typeof MembershipPlan)[keyof typeof MembershipPlan];

export const MembershipPlan = {
  FREE: "FREE",
  PREMIUM: "PREMIUM",
} as const;

export interface Membership {
  plan: MembershipPlan;
  startedAt?: string;
  expiresAt?: string;
  documentsLimit: number;
  aiQueriesLimit: number;
  features: string[];
}

export type UpgradeMembershipBodyPlan =
  (typeof UpgradeMembershipBodyPlan)[keyof typeof UpgradeMembershipBodyPlan];

export const UpgradeMembershipBodyPlan = {
  PREMIUM: "PREMIUM",
} as const;

export interface UpgradeMembershipBody {
  plan: UpgradeMembershipBodyPlan;
  paymentMethod?: string;
}

export type GetLawyersParams = {
  specialization?: string;
  language?: string;
  search?: string;
  minRating?: number;
};

export type GetLawyerAvailabilityParams = {
  date?: string;
};