# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: LexAI - AI Legal Assistant Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/register:
    post:
      operationId: registerUser
      tags: [auth]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterBody"
      responses:
        "201":
          description: User registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
  /auth/login:
    post:
      operationId: loginUser
      tags: [auth]
      summary: Login with email/password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginBody"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
...
  /chat/sessions:
    get:
      operationId: getChatSessions
      tags: [chat]
      summary: Get all chat sessions for current user
      responses:
        "200":
          description: List of chat sessions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ChatSession"
    post:
      operationId: createChatSession
      tags: [chat]
      summary: Create a new chat session
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateChatSessionBody"
      responses:
        "201":
          description: Chat session created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChatSession"
  /chat/sessions/{sessionId}:
    get:
      operationId: getChatSession
      tags: [chat]
      summary: Get chat session with messages
      parameters:
        - name: sessionId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Chat session with messages
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChatSessionDetail"
    delete:
      operationId: deleteChatSession
      tags: [chat]
      summary: Delete a chat session
      parameters:
        - name: sessionId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
  /chat/sessions/{sessionId}/messages:
    post:
      operationId: sendMessage
      tags: [chat]
      summary: Send a message in a chat session
      parameters:
        - name: sessionId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SendMessageBody"
      responses:
        "200":
          description: AI response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChatMessage"

  # DOCUMENTS
...
  /bookings:
    get:
      operationId: getBookings
      tags: [bookings]
      summary: Get user bookings
      responses:
        "200":
          description: Bookings list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Booking"
    post:
      operationId: createBooking
      tags: [bookings]
      summary: Schedule a meeting with a lawyer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateBookingBody"
      responses:
        "201":
          description: Booking created with Meet link
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Booking"
  /bookings/{bookingId}:
    get:
      operationId: getBooking
      tags: [bookings]
      summary: Get booking by ID
      parameters:
        - name: bookingId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Booking detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Booking"
    delete:
      operationId: cancelBooking
      tags: [bookings]
      summary: Cancel a booking
      parameters:
        - name: bookingId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Cancelled
...
components:
  schemas:
...
    RegisterBody:
      type: object
      properties:
        email:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        password:
          type: string
        role:
          type: string
          enum: [NORMAL, LAWYER]
        language:
          type: string
      required: [email, firstName, lastName, password, role]
    LoginBody:
      type: object
      properties:
        email:
          type: string
        password:
          type: string
      required: [email, password]
    AuthResponse:
      type: object
      properties:
        user:
          $ref: "#/components/schemas/User"
        token:
          type: string
      required: [user]
    User:
      type: object
      properties:
        id:
          type: integer
        email:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        role:
          type: string
          enum: [NORMAL, LAWYER]
        avatarUrl:
          type: string
        language:
          type: string
        membershipPlan:
          type: string
          enum: [FREE, PREMIUM]
        createdAt:
          type: string
      required: [id, email, firstName, lastName, role, membershipPlan]
...
    ChatSession:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        documentId:
          type: integer
        lastMessage:
          type: string
        messageCount:
          type: integer
        createdAt:
          type: string
        updatedAt:
          type: string
      required: [id, title, messageCount, createdAt, updatedAt]
    ChatSessionDetail:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        documentId:
          type: integer
        messages:
          type: array
          items:
            $ref: "#/components/schemas/ChatMessage"
        createdAt:
          type: string
        updatedAt:
          type: string
      required: [id, title, messages, createdAt, updatedAt]
    ChatMessage:
      type: object
      properties:
        id:
          type: integer
        sessionId:
          type: integer
        role:
          type: string
          enum: [user, assistant]
        content:
          type: string
        createdAt:
          type: string
      required: [id, sessionId, role, content, createdAt]
    CreateChatSessionBody:
      type: object
      properties:
        title:
          type: string
        documentId:
          type: integer
      required: [title]
    SendMessageBody:
      type: object
      properties:
        content:
          type: string
        documentId:
          type: integer
      required: [content]

    # DOCUMENT SCHEMAS
...
    Booking:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        lawyerId:
          type: integer
        lawyerName:
          type: string
        lawyerAvatarUrl:
          type: string
        scheduledAt:
          type: string
        duration:
          type: integer
        meetLink:
          type: string
        status:
          type: string
          enum: [PENDING, CONFIRMED, CANCELLED, COMPLETED]
        notes:
          type: string
      required: [id, userId, lawyerId, lawyerName, scheduledAt, duration, status]
    CreateBookingBody:
      type: object
      properties:
        lawyerId:
          type: integer
        scheduledAt:
          type: string
        duration:
          type: integer
        notes:
          type: string
      required: [lawyerId, scheduledAt, duration]
...