# lib/api-zod/src/generated/types/booking.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * LexAI - AI Legal Assistant Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { BookingStatus } from "./bookingStatus";

export interface Booking {
  id: number;
  userId: number;
  lawyerId: number;
  lawyerName: string;
  lawyerAvatarUrl?: string;
  scheduledAt: string;
  duration: number;
  meetLink?: string;
  status: BookingStatus;
  notes?: string;
}