# lib/api-zod/src/generated/types/user.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * LexAI - AI Legal Assistant Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { UserMembershipPlan } from "./userMembershipPlan";
import type { UserRole } from "./userRole";

export interface User {
  id: number;
  email: string;
  firstName: string;
  lastName: string;
  role: UserRole;
  avatarUrl?: string;
  language?: string;
  membershipPlan: UserMembershipPlan;
  createdAt?: string;
}